@@ -3,7 +3,7 @@
 
 	<groupId>org.icatproject</groupId>
 	<artifactId>authn.ldap</artifactId>
-	<version>2.0.2-SNAPSHOT</version>
+	<version>3.0.1-SNAPSHOT</version>
 	<packaging>war</packaging>
 	<name>ICAT: authn.ldap</name>
 
@@ -61,11 +61,17 @@
 	</distributionManagement>
 
 	<dependencies>
+		<dependency>
+			<groupId>jakarta.platform</groupId>
+			<artifactId>jakarta.jakartaee-api</artifactId>
+			<version>10.0.0</version>
+			<scope>provided</scope>
+		</dependency>
 
 		<dependency>
 			<groupId>org.icatproject</groupId>
 			<artifactId>icat.authentication</artifactId>
-			<version>4.6.0</version>
+			<version>5.0.0</version>
 		</dependency>
 
 		<dependency>
@@ -103,9 +109,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>org.glassfish</groupId>
-			<artifactId>javax.json</artifactId>
-			<version>1.0.4</version>
+			<groupId>org.eclipse.parsson</groupId>
+			<artifactId>parsson</artifactId>
+			<version>1.1.0</version>
 			<scope>test</scope>
 		</dependency>
 
@@ -199,11 +205,10 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.3</version>
+				<version>3.10.1</version>
 				<configuration>
 					<encoding>${project.build.sourceEncoding}</encoding>
-					<source>1.8</source>
-					<target>1.8</target>
+					<release>11</release>
 				</configuration>
 			</plugin>
 
